# caddyhttp: performance improvement in Regex Match